placementproject/requirements.txt
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==8.3.2
Pint==0.18
psycopg2==2.9.1
pylint==2.7.4
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-decouple==3.5
pytz==2021.3
requests==2.24.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.25.11
Werkzeug==2.0.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
django-ckeditor==6.2.0